requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
tensorflow==2.13.0
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
pyyaml==6.0.1
